Disallowed User to delete guest stock photos

+++ app/assets/javascripts/views/posts/post_show.js
@@ -33,14 +33,21 @@ Vinstagram.Views.PostShow = Backbone.CompositeView.extend({
   },
 
   deletePost: function () {
-    if (this.model.attributes.is_current_user) {
-      this.$el.find('.delete-post').text(" ..deleting.. ")
-      this.model.destroy({
-        success: function () {
-          Vinstagram.posts.reset();
-          Backbone.history.navigate('', {trigger: true});
-        }
-      })
+
+    //THIS IS TEMPORARY CODE TO KEEP PEOPLE FROM DELETEING VINS AND GUESTS
+    //PHOTOS FOR PRESERVING GUEST ACCOUNT
+    if (this.model.id < 25) {
+      alert("You can't delete one of our stock photos but feel free to upload and delete new photos!")
+    } else {
+      if (this.model.attributes.is_current_user) {
+        this.$el.find('.delete-post').text(" ..deleting.. ")
+        this.model.destroy({
+          success: function () {
+            Vinstagram.posts.reset();
+            Backbone.history.navigate('', {trigger: true});
+          }
+        })
+      }
     }
   }
